requirements.txt
flask==2.3.3
SQLAlchemy==2.0.20
pg8000==1.30.1
pytest==7.4.2
pytest-mock==3.11.1
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
alembic==1.12.0
Flask-Migrate==4.0.5
pydantic==2.3.0